+++ src/libs/antares/study/xcast/xcast.cpp
@@ -27,7 +27,7 @@
 
 #include "xcast.h"
 #include "../../logs.h"
-#include "../../inifile.h"
+#include <antares/inifile/inifile.h>
 #include "../study.h"
 #include "../memory-usage.h"
 #include <limits>