+++ src/solver/optimisation/adequacy_patch_csr/set_variable_boundaries.cpp
@@ -110,8 +110,8 @@ void HourlyCSRProblem::setBoundsOnFlows()
 {
     const CORRESPONDANCES_DES_VARIABLES* CorrespondanceVarNativesVarOptim
       = problemeHebdo_->CorrespondanceVarNativesVarOptim[triggeredHour];
-    double* Xmin = problemeAResoudre_.Xmin;
-    double* Xmax = problemeAResoudre_.Xmax;
+    std::vector<double>& Xmin = problemeAResoudre_.Xmin;
+    std::vector<double>& Xmax = problemeAResoudre_.Xmax;
     VALEURS_DE_NTC_ET_RESISTANCES& ValeursDeNTC = problemeHebdo_->ValeursDeNTC[triggeredHour];
 
     // variables bounds: transmissin flows (flow, direct_direct and flow_indirect). For links