+++ src/solver/optimisation/opt_rename_problem.h
@@ -9,16 +9,16 @@ class TargetVectorUpdater
 {
 public:
     TargetVectorUpdater(bool isRenamingProcessed, std::vector<std::string>& target) :
-    target_(target), isRenamingProcessed_(isRenamingProcessed)
+     target_(target), isRenamingProcessed_(isRenamingProcessed)
     {
     }
 
     void UpdateTargetAtIndex(const std::string& full_name, unsigned int index)
     {
-    if (isRenamingProcessed_)
-    {
-          target_[index] = full_name;
-    }
+        if (isRenamingProcessed_)
+        {
+            target_[index] = full_name;
+        }
     }
 
 private:
@@ -34,7 +34,7 @@ public:
     {
     }
 
-    void UpdateTimeStep(int timeStep)
+    void UpdateTimeStep(unsigned int timeStep)
     {
         timeStep_ = timeStep;
     }
@@ -43,6 +43,15 @@ public:
     {
         area_ = area;
     }
+    void SetLinkElementName(unsigned int variable, const std::string& variableType);
+    void SetAreaElementNameHour(unsigned int variable, const std::string& variableType);
+    void SetAreaElementNameWeek(unsigned int variable, const std::string& variableType);
+    void SetAreaElementName(unsigned int variable,
+                            const std::string& variableType,
+                            const std::string& timeStepType);
+    void SetThermalClusterElementName(unsigned int variable,
+                                      const std::string& variableType,
+                                      const std::string& clusterName);
 
     unsigned int timeStep_ = 0;
     std::string origin_;
@@ -55,42 +64,40 @@ class VariableNamer : public Namer
 {
 public:
     using Namer::Namer;
-    void DispatchableProduction(unsigned int var, const std::string& clusterName);
-    void NODU(unsigned int var, const std::string& clusterName);
-    void NumberStoppingDispatchableUnits(unsigned int var, const std::string& clusterName);
-    void NumberStartingDispatchableUnits(unsigned int var, const std::string& clusterName);
-    void NumberBreakingDownDispatchableUnits(unsigned int var, const std::string& clusterName);
-    void NTCDirect(unsigned int var, const std::string& origin, const std::string& destination);
-    void IntercoDirectCost(unsigned int var,
+    void DispatchableProduction(unsigned int variable, const std::string& clusterName);
+    void NODU(unsigned int variable, const std::string& clusterName);
+    void NumberStoppingDispatchableUnits(unsigned int variable, const std::string& clusterName);
+    void NumberStartingDispatchableUnits(unsigned int variable, const std::string& clusterName);
+    void NumberBreakingDownDispatchableUnits(unsigned int variable, const std::string& clusterName);
+    void NTCDirect(unsigned int variable,
+                   const std::string& origin,
+                   const std::string& destination);
+    void IntercoDirectCost(unsigned int variable,
                            const std::string& origin,
                            const std::string& destination);
-    void IntercoIndirectCost(unsigned int var,
+    void IntercoIndirectCost(unsigned int variable,
                              const std::string& origin,
                              const std::string& destination);
-    void ShortTermStorageInjection(unsigned int var, const std::string& shortTermStorageName);
-    void ShortTermStorageWithdrawal(unsigned int var, const std::string& shortTermStorageName);
-    void ShortTermStorageLevel(unsigned int var, const std::string& shortTermStorageName);
-    void HydProd(unsigned int var);
-    void HydProdDown(unsigned int var);
-    void HydProdUp(unsigned int var);
-    void Pumping(unsigned int var);
-    void HydroLevel(unsigned int var);
-    void Overflow(unsigned int var);
-    void FinalStorage(unsigned int var);
-    void LayerStorage(unsigned int var, int layerIndex);
-    void PositiveUnsuppliedEnergy(unsigned int var);
-    void NegativeUnsuppliedEnergy(unsigned int var);
-    void AreaBalance(unsigned int var);
+    void ShortTermStorageInjection(unsigned int variable, const std::string& shortTermStorageName);
+    void ShortTermStorageWithdrawal(unsigned int variable, const std::string& shortTermStorageName);
+    void ShortTermStorageLevel(unsigned int variable, const std::string& shortTermStorageName);
+    void HydProd(unsigned int variable);
+    void HydProdDown(unsigned int variable);
+    void HydProdUp(unsigned int variable);
+    void Pumping(unsigned int variable);
+    void HydroLevel(unsigned int variable);
+    void Overflow(unsigned int variable);
+    void FinalStorage(unsigned int variable);
+    void LayerStorage(unsigned int variable, int layerIndex);
+    void PositiveUnsuppliedEnergy(unsigned int variable);
+    void NegativeUnsuppliedEnergy(unsigned int variable);
+    void AreaBalance(unsigned int variable);
 
 private:
-    void SetThermalClusterVariableName(unsigned int var,
-                                       const std::string& variableType,
-                                       const std::string& clusterName);
-    void SetAreaVariableName(unsigned int var, const std::string& variableType);
-
-    void SetAreaVariableName(unsigned int var, const std::string& variableType, int layerIndex);
-    void SetLinkVariableName(unsigned int var, const std::string& variableType);
-    void SetShortTermStorageVariableName(unsigned int var,
+    void SetAreaVariableName(unsigned int variable,
+                             const std::string& variableType,
+                             int layerIndex);
+    void SetShortTermStorageVariableName(unsigned int variable,
                                          const std::string& variableType,
                                          const std::string& shortTermStorageName);
 };
@@ -100,39 +107,39 @@ class ConstraintNamer : public Namer
 public:
     using Namer::Namer;
 
-    void FlowDissociation(unsigned int numConstraint,
+    void FlowDissociation(unsigned int constraint,
                           const std::string& origin,
                           const std::string& destination);
-    void AreaBalance(unsigned int numConstraint);
-    void FictiveLoads(unsigned int numConstraint);
-    void HydroPower(unsigned int numConstraint);
-    void HydroPowerSmoothingUsingVariationSum(unsigned int numConstraint);
-    void HydroPowerSmoothingUsingVariationMaxDown(unsigned int numConstraint);
-    void HydroPowerSmoothingUsingVariationMaxUp(unsigned int numConstraint);
-    void MinHydroPower(unsigned int numConstraint);
-    void MaxHydroPower(unsigned int numConstraint);
-    void MaxPumping(unsigned int numConstraint);
-    void AreaHydroLevel(unsigned int numConstraint);
-    void FinalStockEquivalent(unsigned int numConstraint);
-    void FinalStockExpression(unsigned int numConstraint);
-    void NbUnitsOutageLessThanNbUnitsStop(unsigned int numConstraint);
-    void NbDispUnitsMinBoundSinceMinUpTime(unsigned int numConstraint);
-    void MinDownTime(unsigned int numConstraint);
-    void PMaxDispatchableGeneration(unsigned int numConstraint);
-    void PMinDispatchableGeneration(unsigned int numConstraint);
-    void ConsistenceNODU(unsigned int numConstraint);
-    void ShortTermStorageLevel(unsigned int numConstraint, const std::string& name);
-    void BindingConstraintHour(unsigned int numConstraint, const std::string& name);
-    void BindingConstraintDay(unsigned int numConstraint, const std::string& name);
-    void BindingConstraintWeek(unsigned int numConstraint, const std::string& name);
+    void AreaBalance(unsigned int constraint);
+    void FictiveLoads(unsigned int constraint);
+    void HydroPower(unsigned int constraint);
+    void HydroPowerSmoothingUsingVariationSum(unsigned int constraint);
+    void HydroPowerSmoothingUsingVariationMaxDown(unsigned int constraint);
+    void HydroPowerSmoothingUsingVariationMaxUp(unsigned int constraint);
+    void MinHydroPower(unsigned int constraint);
+    void MaxHydroPower(unsigned int constraint);
+    void MaxPumping(unsigned int constraint);
+    void AreaHydroLevel(unsigned int constraint);
+    void FinalStockEquivalent(unsigned int constraint);
+    void FinalStockExpression(unsigned int constraint);
+    void NbUnitsOutageLessThanNbUnitsStop(unsigned int constraint, const std::string& clusterName);
+    void NbDispUnitsMinBoundSinceMinUpTime(unsigned int constraint, const std::string& clusterName);
+    void MinDownTime(unsigned int constraint, const std::string& clusterName);
+    void PMaxDispatchableGeneration(unsigned int constraint, const std::string& clusterName);
+    void PMinDispatchableGeneration(unsigned int constraint, const std::string& clusterName);
+    void ConsistenceNODU(unsigned int constraint, const std::string& clusterName);
+    void ShortTermStorageLevel(unsigned int constraint, const std::string& name);
+    void BindingConstraintHour(unsigned int constraint, const std::string& name);
+    void BindingConstraintDay(unsigned int constraint, const std::string& name);
+    void BindingConstraintWeek(unsigned int constraint, const std::string& name);
 
 private:
-    void nameWithTimeGranularity(unsigned int numConstraint,
+    void nameWithTimeGranularity(unsigned int constraint,
                                  const std::string& name,
                                  const std::string& type);
 };
 
-inline std::string TimeIdentifier(int timeStep, const std::string& timeStepType)
+inline std::string TimeIdentifier(unsigned int timeStep, const std::string& timeStepType)
 {
     return timeStepType + "<" + std::to_string(timeStep) + ">";
 }